Fix remaining bugs in mrc.py
The only bad thing is that countries, MPAA, and star ratings
are hard-coded.  If I were to add a new country, MPAA, or star
rating, it would also have to be added in this script too.
Also, I could improve this even more by having the lex analyzer
create a list of abstract syntax lists (list of lists containing
Title object, Year object, StarRating object, Mpaa object, and
Country object).  This is the way I should do it...